# Conflicts:
#	Engine/gfx/ali3dogl.cpp
#	Engine/platform/windows/gfx/ali3dd3d.cpp

+++ Engine/gfx/ali3dogl.cpp
@@ -1071,8 +1071,8 @@ void OGLGraphicsDriver::_renderSprite(const OGLDrawListEntry *drawListEntry,
 
   float width = bmpToDraw->GetWidthToRender();
   float height = bmpToDraw->GetHeightToRender();
-  float xProportion = (float)width / (float)bmpToDraw->_width;
-  float yProportion = (float)height / (float)bmpToDraw->_height;
+  float xProportion = width / (float)bmpToDraw->_width;
+  float yProportion = height / (float)bmpToDraw->_height;
   int drawAtX = drawListEntry->x;
   int drawAtY = drawListEntry->y;
 
@@ -1087,14 +1087,14 @@ void OGLGraphicsDriver::_renderSprite(const OGLDrawListEntry *drawListEntry,
       xOffs = (bmpToDraw->_width - (txdata->_tiles[ti].x + txdata->_tiles[ti].width)) * xProportion;
     else
       xOffs = txdata->_tiles[ti].x * xProportion;
-    int thisX = drawAtX + xOffs;
-    int thisY = drawAtY + yOffs;
-    thisX = (-(surface_size.Width / 2)) + thisX;
-    thisY = (surface_size.Height / 2) - thisY;
+    float thisX = drawAtX + xOffs;
+    float thisY = drawAtY + yOffs;
+    thisX = (-(surface_size.Width / 2.0f)) + thisX;
+    thisY = (surface_size.Height / 2.0f) - thisY;
 
     //Setup translation and scaling matrices
-    float widthToScale = (float)width;
-    float heightToScale = (float)height;
+    float widthToScale = width;
+    float heightToScale = height;
     if (bmpToDraw->_flipped)
     {
       // The usual transform changes 0..1 into 0..width
@@ -1120,7 +1120,7 @@ void OGLGraphicsDriver::_renderSprite(const OGLDrawListEntry *drawListEntry,
     // Global batch transform
     transform = transform * matGlobal;
     // Self sprite transform (first scale, then rotate and then translate, reversed)
-    transform = glmex::transform2d(transform, (float)thisX, (float)thisY, widthToScale, heightToScale,
+    transform = glmex::transform2d(transform, thisX, thisY, widthToScale, heightToScale,
         rotZ, pivotX, pivotY);
 
     glUniformMatrix4fv(program.MVPMatrix, 1, GL_FALSE, glm::value_ptr(transform));
