+++ Engine/platform/windows/gfx/ali3dd3d.cpp
@@ -1054,12 +1054,12 @@ void D3DGraphicsDriver::_renderSprite(const D3DDrawListEntry *drawListEntry, con
       xOffs = txdata->_tiles[ti].x * xProportion;
     float thisX = drawAtX + xOffs;
     float thisY = drawAtY + yOffs;
-    thisX = (-(surface_size.Width / 2)) + thisX;
-    thisY = (surface_size.Height / 2) - thisY;
+    thisX = (-(surface_size.Width / 2.0f)) + thisX;
+    thisY = (surface_size.Height / 2.0f) - thisY;
 
     //Setup translation and scaling matrices
-    float widthToScale = (float)width;
-    float heightToScale = (float)height;
+    float widthToScale = width;
+    float heightToScale = height;
     if (bmpToDraw->_flipped)
     {
       // The usual transform changes 0..1 into 0..width
@@ -1077,7 +1077,7 @@ void D3DGraphicsDriver::_renderSprite(const D3DDrawListEntry *drawListEntry, con
 
     // Self sprite transform (first scale, then rotate and then translate, reversed)
     glm::mat4 transform = glmex::make_transform2d(
-        (float)thisX - _pixelRenderXOffset, (float)thisY + _pixelRenderYOffset, widthToScale, heightToScale,
+        thisX - _pixelRenderXOffset, thisY + _pixelRenderYOffset, widthToScale, heightToScale,
         rotZ, pivotX, pivotY);
     // Global batch transform
     transform = matGlobal * transform;
@@ -1193,7 +1193,7 @@ void D3DGraphicsDriver::_renderFromTexture()
     float drawAtY = _srcRect.GetHeight() / 2;
 
     glm::mat4 transform = glmex::make_transform2d(
-        (float)drawAtX - _pixelRenderXOffset, (float)drawAtY + _pixelRenderYOffset, width, height, 0.f);
+        drawAtX - _pixelRenderXOffset, drawAtY + _pixelRenderYOffset, width, height, 0.f);
 
     direct3ddevice->SetTransform(D3DTS_WORLD, (D3DMATRIX*)glm::value_ptr(transform));
